Refactor fetch_last_entries function and add display_entries function
This commit refactors the fetch_last_entries function to accept a limit parameter, allowing for fetching a variable number of entries. It also adds a new function, display_entries, which takes a list of entries and displays them in a formatted table. These changes improve the flexibility and readability of the code.